Update README.md
Near the start:
verless (pronounced like serverless) is a Static Site Generator 
designed for Markdown-based content with ** a** focus on simplicity and performance.

There probably should be an "a" there so I've put one in. It could work without one but only in very specific
situations in my opinion.

Near the end, before the documentation:

"You're now good to create your first content!"
(This is ok, the main thing I'd change is "your first content"

Better: You are now ready to create your first blog post!

Or:: You are now ready to start creating content!